created a dynamic webpage page, where data is dynamicly places. checked, error. now making pages is easier and faster.
still a lot of work. need to set in the right key, place the variable in the right type. make sure the naming convention is the same in the backend. but that should be it. now I don't have to copy paste code over and over again, AND test it, now everything will be checked wiht the DTO or the Backend if it passses

in the backend the code is checked. and a message system, cauhgt in the frontend, WILL find it and use it. make sure the message has the correct key word as the first word for it to work