you cna now clikc to book a room. however. in here you also need to decide the begin and start date.
Begin and Start date could Have been provided EARLIER. this is going to be complicated. because that data i send TO the page in ALLROOMS. shows all the dates you COULD NOT get through....

everything works now.

Date will be given and the backend gives all the beginDate and Start dates from every Room. a normal hotel might also give that, and I can't restrict the data, its required.

even then, someone could make a code that continuesly asks the hotel for a certain date, and if it returns back "yes" so anyway. this seems faster and with less hickups.