setting up roles. should be easy
each role is their own unique entity. added to it will be the webites its allowed to visit. in short.

at the refresh page, it will filter each website it can have. it will send back the names of the websites based on the roles one has. unless one is admit. then they're allowed in.

each HTTP request is already protected by the owner, you cannot change the owner unless you are either the hotel owner, or, if its a user case, the user.

dame here, the Admit could see things, but maybe not change anything. unless its required, in that cas,e an IT function is added to the backend, resulting in them being able to pull data. this is for now, not important.

at the start of creating a user, one setsup their roles. Admit, Owner, User.

only admit is allowed to add webpages to each role. the naming is given in the function of the refresh rate. this is used in the dropdown, and its corresponding role. this way, one can add rights to webpages, or, take them away.