added role check system
@Post(`CreateHotel`)
  @Permissions('CreateHotel')
  @UseGuards(CheckRolesGuard)

by using my permissions, you now chekc the path. if the path is NOT added to any of the roles you have. then that means you are not allowed in there

in the Role Setup, admins can add paths to roles. or setup entire new roles with new paths.

right now, the getALLPaths is not getting all the paths in my backend, surely, there must be a way?